requirements/common.txt
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
pip>=20.0.2
pytz==2019.3
psycopg2==2.8.4